Fixed compile warnings in clang
- Fixed a -Wshorten-64-to-32 warning in case there's no thread-local-storage available
- Disabled -Wcast-align for the time being

+++ src/common/thread.c
@@ -263,7 +263,7 @@ int rathread_get_tid(void) {
 	#ifdef WIN32
 		return (int)GetCurrentThreadId();
 	#else
-		return (intptr_t)pthread_self();
+		return (int)pthread_self();
 	#endif
 #endif
